refactor: 아키텍쳐 설계를 변경한다
- 상태관리를 위해 flux 패턴을 도입한다
- 구매 금액 입력란과 로또 상세보기 화면을 컴포넌트화 한다
