+++ src/js/components/index.js
@@ -1,0 +1,2 @@
+import "./cashContainer.js";
+import "./lottoDetailContainer.js";
