+++ src/js/constants/actionType.js
@@ -1,0 +1,5 @@
+const ACTION_TYPE = {
+  LOTTOS: "lottos",
+};
+
+export default ACTION_TYPE;
