+++ src/js/constants/index.js
@@ -1,5 +1,6 @@
 import JS_SELECTOR from "./jsSelector.js";
 import ALERT_MESSAGE from "./alertMessage.js";
 import CLASSNAME from "./className.js";
+import ACTION_TYPE from "./actionType.js";
 
-export { JS_SELECTOR, ALERT_MESSAGE, CLASSNAME };
+export { ACTION_TYPE, JS_SELECTOR, ALERT_MESSAGE, CLASSNAME };
