+++ src/js/models/Lotto.js
@@ -1,5 +1,5 @@
 export default class Lotto {
   constructor(numbers) {
-    this.numbers = numbers;
+    this.numbers = [...numbers].sort((a, b) => a - b);
   }
 }
