+++ src/js/models/index.js
@@ -1,0 +1,3 @@
+import Lotto from "./Lotto.js";
+
+export { Lotto };
